Changed code sending behavior
Before only files were saved and then code was executed. Now if we send
only a single line or the selection if less than 8192 chars long we send
only the code filtered with comments removed.
If we send a whole file we just run the do "`file"'command.
If the file an unsaved new file or we have a large selection we now write
the temp file and then execute it in Stata.