Add option to send Stata code to VS Code terminal
Sometimes it is helpful to send the code directly to the VS Code terminal windows rather than a Stata application.

This commits adds support for the VS Code terminal as a target for StataRun.

This functionality is toggled through a new boolean option.

In case there is no active terminal window, StataRun will create a new window and open a new Stata session.